Backend/webapp/requirements.txt
Django
Pillow
annoy
django-cors-headers
django-tagging
djangorestframework
pandas
scipy
psycopg2
whitenoise
gunicorn